spark_codegen/requirements.txt
bigquery
pandas
google-cloud-aiplatform
db-dtypes